feature: navigating to /api/users or /api/messages returns error

+++ backend/routes/message.js
@@ -2,24 +2,45 @@ const express = require('express');
 const router = express.Router();
 const Message = require('../models/message.js');
 const cors = require('cors');
+const jwt = require('jsonwebtoken');
 
 router.use(cors());
 
-// SOCKETIOBS
+const getTokenFrom = (req) => {
+  const authorization = req.get('authorization');
+
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7);
+  }
+  return null;
+};
 
 /* GET ALL MESSAGES */
 router.get('/', async (req, res, next) => {
+  const token = getTokenFrom(req);
+  if (!token) {
+    res.sendStatus(401);
+  }
+
+  const decodedToken = jwt.verify(token, 'SALAISUUS');
+  if (!decodedToken.id) {
+    res.status(401).json({ error: 'token missing or invalid' });
+  }
   await Message.find(function (err, products) {
     if (err) return next(err);
-    // req.io.on('connection', (socket) => {
-    //   req.io.emit('messages', products);
-    // });
     res.json(products);
   });
 });
 
 /* GET SINGLE MESSAGE BY ID */
 router.get('/:id', async (req, res, next) => {
+  const token = getTokenFrom(req);
+
+  const decodedToken = jwt.verify(token, 'SALAISUUS');
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
   await Message.findById(req.params.id, function (err, post) {
     if (err) return next(err);
     res.json(post);
@@ -28,6 +49,13 @@ router.get('/:id', async (req, res, next) => {
 
 /* SAVE MESSAGE */
 router.post('/', async (req, res, next) => {
+  const token = getTokenFrom(req);
+
+  const decodedToken = jwt.verify(token, 'SALAISUUS');
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
   await Message.create(req.body, function (err, post) {
     if (err) return next(err);
     res.json(post);
@@ -36,6 +64,13 @@ router.post('/', async (req, res, next) => {
 
 /* UPDATE MESSAGE */
 router.put('/:id', async (req, res, next) => {
+  const token = getTokenFrom(req);
+
+  const decodedToken = jwt.verify(token, 'SALAISUUS');
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
   await Message.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -48,6 +83,13 @@ router.put('/:id', async (req, res, next) => {
 
 /* DELETE MESSAGE */
 router.delete('/:id', async (req, res, next) => {
+  const token = getTokenFrom(req);
+
+  const decodedToken = jwt.verify(token, 'SALAISUUS');
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
   await Message.findByIdAndRemove(
     req.params.id,
     req.body,
